pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nanocoderagent"
version = "0.1.0"
description = "Minimal AI coding agent (~1300 LoC) inspired by Claude Code. Works with any LLM."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Yufeng He", email = "40085740+he-yufeng@users.noreply.github.com" },
]
keywords = ["ai", "agent", "coding", "cli", "llm", "claude-code"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development",
]
dependencies = [
    "openai>=1.0",
    "rich>=13.0",
    "prompt_toolkit>=3.0",
    "python-dotenv>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["nanocoder"]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
nanocoder = "nanocoder.cli:main"

[project.urls]
Homepage = "https://github.com/he-yufeng/NanoCoder"
Repository = "https://github.com/he-yufeng/NanoCoder"
Issues = "https://github.com/he-yufeng/NanoCoder/issues"
